extract ancillary data decoding to helper function
